organized audio analysis code
