+++ src/lib/audio-analysis/track-features.ts
@@ -1,0 +1,15 @@
+export interface TrackFeatures {
+  id: string;
+  href: string;
+  acousticness: number;
+  danceability: number;
+  energy: number;
+  instrumentalness: number;
+  key: number;
+  liveness: number;
+  loudness: number;
+  mode: number;
+  speechiness: number;
+  tempo: number;
+  valence: number;
+}
